all ready commit

+++ mess.py
@@ -1,36 +1,35 @@
 import telebot
 import os
 import time
-import json
-import os
+import chardet
 
-token = '6419893616:AAG-tbu524ZN7IGIulbJA_ZxNLykdaJWeU0'
-bot = telebot.TeleBot(token, parse_mode = None)
-users_id = [412850740]
+token = '5814873337:AAFmEDxaPRXmg8w1HQ4FTiNB1U5l8pgtFgE'
+bot = telebot.TeleBot(token, parse_mode=None)
+users_id = [412850740]  # Замените на свой список ID пользователей
 
 def send_message(txt_file):
-    if os.stat(txt_file).st_size > 0:
-        with open(txt_file, 'r') as fr:
-            mess = json.load(fr)
+    if os.path.exists(txt_file) and os.stat(txt_file).st_size > 0:
+        encoding = detect_encoding(txt_file)
+        with open(txt_file, 'r', encoding=encoding) as fr:
+            mess = fr.read()
         for user in users_id:
             try:
                 bot.send_message(user, mess)
-            except:
-                continue
+            except Exception as e:
+                print(f"Error sending message to user {user}: {e}")
         with open(txt_file, 'w') as fw:
             pass
 
+def detect_encoding(file_path):
+    with open(file_path, 'rb') as file:
+        result = chardet.detect(file.read())
+    return result['encoding']
+
 while True:
     time.sleep(1)
-    try:
-        send_message('usd.txt')
-        send_message('eur.txt')
-        send_message('cny.txt')
+    send_message('sig_proc.txt')
 
+bot.polling(none_stop=True)
 
-    except json.decoder.JSONDecodeError as err:
-        bot.send_message(users_id[0], 'Ошибка: ' + str(err))
-
-bot.infinity_polling()
